Moves NamedCommand under Plugins
