+++ src/Plugins/NamedCommand/NamedCommandExtractor.php
@@ -1,6 +1,8 @@
 <?php
 
-namespace League\Tactician\Handler\CommandNameExtractor;
+namespace League\Tactician\Plugins\NamedCommand;
+
+use League\Tactician\Handler\CommandNameExtractor\CommandNameExtractor;
 
 /**
  * Extract the name from a NamedCommand
